badge component moved
