+++ pages/assessments/dashboard/[badges]/index.tsx
@@ -2,82 +2,13 @@ import React from 'react';
 import { useState, useEffect } from 'react';
 import { useRouter } from 'next/router';
 import Image from 'next/image';
-import Link from 'next/link';
 import MainLayout from '../../../../components/Layout/MainLayout';
+import BadgeComponent from '@modules/assessment/component/Badges/BadgeComponent';
 
-interface BadgeComponentProps {
-  locked: boolean;
-  imageSrc: string;
-  imageAlt: string;
-  title: string;
-  description: string;
-  earnedDate: string;
-  isLocked: boolean;
-  href: string;
-  badgelabel: string;
-  as: string;
-}
-
-type AssessmentDetails = {
-  assessment_id: number;
-  skill_id: number;
-  title?: string;
-  description: string;
-  duration_minutes: string;
-  status: string;
-  start_date: Date;
-  end_date: Date;
-};
-
-const BadgeComponent: React.FC<BadgeComponentProps> = ({
-  imageSrc,
-  imageAlt,
-  title,
-  description,
-  earnedDate,
-  isLocked,
-  badgelabel,
-}) => {
-  const [result, setResult] = React.useState<AssessmentDetails>();
-  const id = '02cac250-ccbf-409f-9c67-ecbbdb5bc31e';
-
-  return (
-    <Link
-      href={`/assessments/dashboard/badge/$[id]`}
-      as={`/assessments/dashboard/badges/badge/${badgelabel}`}
-      passHref
-      className="badgecomponent w-[330px] h-[330px] sm:w-[236px] sm:h-[236px] lg:min-w-[300px] lg:min-h-[300px] xl:min-w-[330px] xl:min-h-[330px] p-[16px] flex flex-col items-center border border-neutral-200 rounded-[8px] gap-[12px] relative overflow-hidden"
-    >
-      {/* {isLocked && (
-          <div className="badgelockoverlay absolute w-full h-full top-0 bg-opacity-60 bg-black flex items-center justify-center z-10 ">
-          <div className="badgelockimage w-[139px] sm:w-[90px] lg:w-[139px] lg:h-[139px]  h-[139px] sm:h-[90px] relative">
-            <Image
-              src="/assets/images/badges/badgelock.png"
-              fill={true}
-              alt="expert reward badge image"
-              priority={true}
-              />
-          </div>
-        </div>
-      )} */}
-      <div className="badgerewardimage w-[160px] h-[128px] relative">
-        <Image src={imageSrc} fill={true} alt={imageAlt} style={{ objectFit: 'contain' }} />
-      </div>
-      <div className="flex flex-col gap-[12px]">
-        <h1 className="text-center text-[22px] sm:text-[16px] lg:text-[22px] font-[600]">{title}</h1>
-        <p className="text-center block sm:hidden lg:block text-[14px] font-[400] leading-[20px] tracking-wide w-full">
-          {description}
-        </p>
-        <p className="text-center block sm:hidden lg:block font-[700] leading-[20px] tracking-wider">{earnedDate}</p>
-      </div>
-    </Link>
-  );
-};
-
-const Earnedbadges: React.FC<BadgeComponentProps> = ({ locked }) => {
+const Earnedbadges: React.FC = () => {
   const [isLoading, setIsLoading] = useState<boolean>(true);
   const router = useRouter();
-  const badgelabel = router.query?.id;
+  const badgelabel = router.query?.badge;
   console.log(badgelabel);
 
   const [badgeData, setBadgeData] = useState(null) as any;
@@ -103,7 +34,7 @@ const Earnedbadges: React.FC<BadgeComponentProps> = ({ locked }) => {
     };
 
     fetchData();
-  }, []);
+  }, [badgeData, badgelabel]);
 
   return (
     <MainLayout activePage="marketplace" showDashboardSidebar={false} showFooter={true} showTopbar={true}>
@@ -192,7 +123,7 @@ const Earnedbadges: React.FC<BadgeComponentProps> = ({ locked }) => {
               earnedDate="Earned on: Sept 12, 2023"
               isLocked={false}
               locked={true}
-              href="/badges/badge/[badge]"
+              // href="/badges/badge/[badge]"
               badgelabel="beginner"
               as={''}
             />
@@ -206,7 +137,7 @@ const Earnedbadges: React.FC<BadgeComponentProps> = ({ locked }) => {
               isLocked={true}
               locked={true}
               badgelabel="intermediate"
-              href="/badges/badge/[badge]"
+              // href="/badges/badge/[badge]"
               as={''}
             />
 
@@ -221,7 +152,7 @@ const Earnedbadges: React.FC<BadgeComponentProps> = ({ locked }) => {
               isLocked={true}
               locked={true}
               badgelabel="expert"
-              href="/badges/badge/[badge]"
+              // href="/badges/badge/[badge]"
               as={''}
             />
           </div>
